backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
mysql-connector-python==9.5.0
